requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pyjwt==2.6.0
python-multipart==0.0.6
